SHRL-30 moved ListLogHandler to src/test/java
